requirements.txt
asgiref==3.11.0
bidict==0.23.1
blinker==1.9.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
Django==6.0.1
django-unfold==0.76.0
Flask==3.1.2
Flask-SocketIO==5.6.0
h11==0.16.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
pillow==12.1.0
pyirsdk==1.3.5
python-engineio==4.13.0
python-socketio==5.16.0
PyYAML==6.0.3
requests==2.32.3
simple-websocket==1.1.0
sqlparse==0.5.5
tzdata==2025.3
urllib3==2.6.2
Werkzeug==3.1.4
wsproto==1.3.2
gunicorn
whitenoise
dj-database-url
psycopg2-binary
